Fix a nested property assignment mutation issue
When evaluating foo[xxx] = 'quux', evaluation of 'xxx' may mutate the
register bound variable 'foo'.  So even if it seems we could safely
assume 'foo' maps to a register (without emitting any code) we need to
copy it to a temporary unless we know 'xxx' doesn't mutate it as a side
effect.  Duktape doesn't currently have the compiler state to do such a
check so we must assume the worst.  Even with a compiler internal
representation it's difficult to prove 'xxx' is safe if it involves any
side effects; e.g. foo[bar.quux] = 'quux' is potentially unsafe because
'bar.quux' may be an accessor and write to 'foo'.